+++ docs/Whats-New.md
@@ -56,6 +56,7 @@ This page lists the history of changes across stable Vinifera releases and also 
 New:
 - Implement voxel light customization (by ZivDero)
 - Implement an option to disable Tiberium storage (by Rampastring)
+- Implement TActionClass extensions, port ts-patches actions (by ZivDero, Rampastring)
 
 Vinifera fixes:
 - Fix unit placement in non-TS Client builds of Vinifera (by ZivDero)